refactor: move fetch helpers into script
